// src/generated/prisma/models/MainCalculation.ts
export type MainCalculationWhereInput = {
  AND?: Prisma.MainCalculationWhereInput | Prisma.MainCalculationWhereInput[]
  OR?: Prisma.MainCalculationWhereInput[]
  NOT?: Prisma.MainCalculationWhereInput | Prisma.MainCalculationWhereInput[]
  id?: Prisma.StringFilter<"MainCalculation"> | string
  calculationDate?: Prisma.DateTimeFilter<"MainCalculation"> | Date | string
  totalTabil?: Prisma.IntFilter<"MainCalculation"> | number
  dailyCalculationId?: Prisma.StringFilter<"MainCalculation"> | string
  interest?: Prisma.IntFilter<"MainCalculation"> | number
  bandak?: Prisma.IntFilter<"MainCalculation"> | number
  jinisChara?: Prisma.IntFilter<"MainCalculation"> | number
  cash?: Prisma.IntFilter<"MainCalculation"> | number
  leftTotal?: Prisma.IntFilter<"MainCalculation"> | number
  rightTotal?: Prisma.IntFilter<"MainCalculation"> | number
  difference?: Prisma.IntFilter<"MainCalculation"> | number
  balanceStatus?: Prisma.EnumDailyCalculationBalanceStatusFilter<"MainCalculation"> | $Enums.DailyCalculationBalanceStatus
  recordStatus?: Prisma.EnumMainCalculationRecordStatusFilter<"MainCalculation"> | $Enums.MainCalculationRecordStatus
  finalizedAt?: Prisma.DateTimeNullableFilter<"MainCalculation"> | Date | string | null
  createdById?: Prisma.StringFilter<"MainCalculation"> | string
  createdAt?: Prisma.DateTimeFilter<"MainCalculation"> | Date | string
  updatedAt?: Prisma.DateTimeFilter<"MainCalculation"> | Date | string
  lastEditedById?: Prisma.StringNullableFilter<"MainCalculation"> | string | null
  lastEditedAt?: Prisma.DateTimeNullableFilter<"MainCalculation"> | Date | string | null
  dailyCalculation?: Prisma.XOR<Prisma.DailyCalculationScalarRelationFilter, Prisma.DailyCalculationWhereInput>
  createdBy?: Prisma.XOR<Prisma.UserScalarRelationFilter, Prisma.UserWhereInput>
  lastEditedBy?: Prisma.XOR<Prisma.UserNullableScalarRelationFilter, Prisma.UserWhereInput> | null
}